# gdrive-aggregator-be/docs/swagger.yaml
basePath: /
definitions:
  handler.BatchDeleteRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  model.BaseResponse:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  model.ErrorResponse:
    properties:
      error:
        example: detail error
        type: string
      message:
        example: Terjadi kesalahan
        type: string
      success:
        example: false
        type: boolean
    type: object
  model.FileRecord:
    properties:
      account_email:
        type: string
      account_id:
        type: integer
      created_at:
        type: string
      drive_file_id:
        type: string
      icon_link:
        type: string
      id:
        type: integer
      is_folder:
        type: boolean
      md5_checksum:
        type: string
      mime_type:
        type: string
      name:
        type: string
      parent_id:
        type: string
      size:
        description: bytes
        type: integer
      updated_at:
        type: string
      web_view_link:
        type: string
    type: object
  model.FileUploadResponseDTO:
    properties:
      account_email:
        example: user@gmail.com
        type: string
      account_id:
        example: 1
        type: integer
      drive_file_id:
        example: 1a2b3c4d...
        type: string
      file_id:
        example: 1
        type: integer
      mime_type:
        example: application/pdf
        type: string
      name:
        example: document.pdf
        type: string
      size:
        example: 1048576
        type: integer
      size_human:
        example: 1.00 MB
        type: string
      web_view_link:
        example: https://drive.google.com/file/d/...
        type: string
    type: object
...
info:
  contact:
    name: Google Drive Aggregator Support
  description: |-
    REST API untuk mengagregasi beberapa akun Google Drive ke dalam satu storage pool terpusat.
    Mendukung smart allocation (otomatis pilih akun dengan sisa kuota terbesar), streaming download, dan manajemen multi-akun.
  title: Google Drive Aggregator API
  version: "1.0"
paths:
...
  /api/v1/files:
    get:
      consumes:
      - application/json
      description: Menampilkan metadata file dari seluruh akun Google Drive yang teragregasi
      parameters:
      - description: Filter berdasarkan ID akun
        in: query
        name: account_id
        type: integer
      - description: Cari berdasarkan nama file
        in: query
        name: search
        type: string
      - description: Jumlah per halaman (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.FileRecord'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Daftar Semua File
      tags:
      - Files
  /api/v1/files/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus file dari Google Drive dan database lokal
      parameters:
      - description: ID file di database
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Hapus File
      tags:
      - Files
  /api/v1/files/batch-delete:
    post:
      consumes:
      - application/json
      description: Menghapus kumpulan file terpilih dari Google Drive dan database
        lokal
      parameters:
      - description: Daftar ID file yang ingin dihapus
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BatchDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.BaseResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Hapus Beberapa File Sekaligus
      tags:
      - Files
  /api/v1/files/download/{id}:
    get:
      description: Mengunduh file langsung dari Google Drive akun bersangkutan via
        streaming
      parameters:
      - description: ID file di database
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File stream
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Stream Download File
      tags:
      - Files
  /api/v1/files/sync:
    post:
      consumes:
      - application/json
      description: Memindai dan mengimpor file yang sudah ada di akun Google Drive
        ke database lokal
      parameters:
      - description: ID Akun Google Drive (opsional, jika kosong sinkronisasi semua
          akun)
        in: query
        name: account_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.BaseResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Sinkronisasi/Impor File dari Google Drive
      tags:
      - Files
  /api/v1/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Mengunggah file biner. Jika account_id tidak diisi, sistem otomatis
        memilih akun dengan sisa kuota terbesar (*Smart Storage Allocation*).
      parameters:
      - description: Pilih file yang ingin diunggah
        in: formData
        name: file
        required: true
        type: file
      - description: 'ID Akun tujuan (opsional, default: auto-pilih akun paling lega)'
        in: formData
        name: account_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.FileUploadResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Upload File ke Google Drive
      tags:
      - Files
...
swagger: "2.0"